requirements.txt
streamlit
tensorflow>=2.16
numpy
pillow
requests
gdown
h5py


